use set_timecourses in the TimeSeries init function
To preprocess the data of the micro lesions experiments I had to construct a new TimeSeries object and set its time course in the constructor. I assumed it will automatically reshape the data also when I give them in the trialShaped2D format. I think it should be done in the constructor.

also add some documentation of the TimeSeries Object