fix the mean bug described in the google doc
the bug was introduced in 3233a01a67ff6eb8d where I started to use set_timecourses in the TimeSeries constructor for correct reshaping. This sometimes added a empty dimension to a TimeSeries object which messed up the filtering of bf.SelectTrials. This behavior is fixed by adding squeeze in set_timeseries

Additionally the first dimension is chose explicitly in the SelectTrials filter. By this the error would have been much easier to spot